Refactor and final settings edit.
