+++ pom.xml
@@ -9,7 +9,7 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.pet</groupId>
-	<artifactId>student-managment-system</artifactId>
+	<artifactId>student-management-system</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>student-managment-system</name>
 	<description>Student Managment System using Spring Boot and Thymeleaf</description>
