+++ src/main/java/com/pet/sms/StudentManagementSystemApplication.java
@@ -1,0 +1,23 @@
+package com.pet.sms;
+
+import com.pet.sms.entity.Student;
+import com.pet.sms.repository.StudentRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StudentManagementSystemApplication implements CommandLineRunner {
+
+	public static void main(String[] args) {
+		SpringApplication.run(StudentManagementSystemApplication.class, args);
+	}
+
+	@Autowired
+	private StudentRepository studentRepository;
+	@Override
+	public void run(String... args) throws Exception {
+
+	}
+}
