+++ src/main/java/com/pet/sms/entity/Student.java
@@ -1,4 +1,4 @@
-package com.pet.studentmanagmentsystem.entity;
+package com.pet.sms.entity;
 
 import jakarta.persistence.*;
