+++ src/main/java/com/pet/sms/service/StudentService.java
@@ -1,0 +1,9 @@
+package com.pet.sms.service;
+
+import com.pet.sms.entity.Student;
+
+import java.util.List;
+
+public interface StudentService {
+    List<Student> getAllStudent();
+}
