+++ src/main/java/com/pet/sms/service/impl/StudentServiceImpl.java
@@ -1,0 +1,23 @@
+package com.pet.sms.service.impl;
+
+import com.pet.sms.entity.Student;
+import com.pet.sms.repository.StudentRepository;
+import com.pet.sms.service.StudentService;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class StudentServiceImpl implements StudentService {
+    private StudentRepository studentRepository;
+
+    public StudentServiceImpl(StudentRepository studentRepository) {
+        super();
+        this.studentRepository = studentRepository;
+    }
+
+    @Override
+    public List<Student> getAllStudent() {
+        return studentRepository.findAll();
+    }
+}
