--- src/main/java/com/pet/studentmanagmentsystem/StudentManagmentSystemApplication.java
@@ -1,13 +1,0 @@
-package com.pet.studentmanagmentsystem;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StudentManagmentSystemApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(StudentManagmentSystemApplication.class, args);
-	}
-
-}
