--- src/main/java/com/pet/studentmanagmentsystem/controller/StudentController.java
@@ -1,7 +1,0 @@
-package com.pet.studentmanagmentsystem.controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class StudentController {
-}
