--- src/main/java/com/pet/studentmanagmentsystem/service/StudentService.java
@@ -1,4 +1,0 @@
-package com.pet.studentmanagmentsystem.service;
-
-public interface StudentService {
-}
