--- src/main/java/com/pet/studentmanagmentsystem/service/impl/StudentServiceImpl.java
@@ -1,7 +1,0 @@
-package com.pet.studentmanagmentsystem.service.impl;
-
-import com.pet.studentmanagmentsystem.service.StudentService;
-
-public class StudentServiceImpl implements StudentService {
-
-}
