+++ src/main/resources/application.properties
@@ -3,7 +3,7 @@ spring.datasource.username=root
 spring.datasource.password=root
 
 #Hibernate
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
 
 #Hibernate auto ddl
 spring.jpa.hibernate.ddl-auto=update
