+++ src/test/java/com/pet/sms/StudentManagementSystemApplicationTests.java
@@ -1,10 +1,10 @@
-package com.pet.studentmanagmentsystem;
+package com.pet.sms;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class StudentManagmentSystemApplicationTests {
+class StudentManagementSystemApplicationTests {
 
 	@Test
 	void contextLoads() {
